PW-32: introduce error boundary component
