+++ src/components/search/index.tsx
@@ -1,19 +1,29 @@
 import React from 'react';
-import {
-  Wrapper,
-  Title,
-  Text,
-  InputSection,
-  InputText,
-  InputIconSection,
-  SubmitButton,
-  FormSearch,
-} from './Search.css';
-import Kod from '../../assets/kod.svg';
-import Microphone from '../../assets/microphone.svg';
+import styled from 'styled-components';
 import { SearchResultsList } from './SearchResultsList';
 import { IProductData } from '../../domain/products';
-import {ButtonColor} from './../buttons/Button';
+import { SearchInput } from './SearchInput';
+import ErrorBoundary from '../../utils/error-boundary';
+import { Device } from '../../styles/theme';
+
+export const Wrapper = styled.div`
+  width: 100%;
+  padding: 260px 0 70px 0;
+  position: relative;
+
+  @media ${Device.mobile} {
+    padding: 100px 0 20px 0;
+    display: flex;
+    align-items: center;
+    flex-direction: column;
+  }
+
+  > div {
+    @media ${Device.desktop} {
+      width: 90%;
+    }
+  }
+`;
 
 interface ISearchContainer {
   searchResults: IProductData[];
@@ -33,49 +43,22 @@ export const SearchContainer: React.FC<ISearchContainer> = ({
   onLoadMore,
   onSelect,
 }) => {
-  const [phrase, setPhrase] = React.useState<string>('');
-  const hasPhrase = !!phrase && phrase.length > 0;
-
-  const handlePhraseChange = (e: React.ChangeEvent<HTMLInputElement>) => setPhrase(e.currentTarget.value);
-  const handleSearch = (e: React.MouseEventHandler<HTMLButtonElement>) => onSearch(phrase);
   const handleLoad = () => onLoadMore();
 
   return (
-    <Wrapper>
-      <div>
-        <Title>Sprawdź informacje o produkcie</Title>
-        <Text>
-          Wpisz tekst, podyktuj lub zeskanuj kod
-          <br />
-          Nie znasz kodu?{' '}
-          <a target="blank" href="https://pl.openfoodfacts.org/">
-            Znajdź go w bazie
-          </a>
-        </Text>
-        <FormSearch>
-          <InputSection>
-            <InputText placeholder="Nazwa produktu/producent/kod EAN" type="text" onChange={handlePhraseChange} />
-            <InputIconSection>
-              <div>
-                <img src={Kod} />
-              </div>
-              <div>
-                <img src={Microphone} />
-              </div>
-            </InputIconSection>
-          </InputSection>
-          <SubmitButton label='Sprawdź' color={ButtonColor.Red} disabled={!hasPhrase} onClick={handleSearch}/>
-        </FormSearch>
-      </div>
-      {searchResults && (
-        <SearchResultsList
-          results={searchResults}
-          token={token}
-          isLoading={isLoading}
-          onLoadMore={handleLoad}
-          onSelect={onSelect}
-        />
-      )}
-    </Wrapper>
+    <ErrorBoundary scope="search-container">
+      <Wrapper>
+        <SearchInput onSearch={onSearch} />
+        {searchResults && (
+          <SearchResultsList
+            results={searchResults}
+            token={token}
+            isLoading={isLoading}
+            onLoadMore={handleLoad}
+            onSelect={onSelect}
+          />
+        )}
+      </Wrapper>
+    </ErrorBoundary>
   );
 };
